change gethandle return type
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@4493 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ common/platform/unix/SysFile.hpp
@@ -126,7 +126,7 @@ public:
     inline void clearErrors() { errInfo = 0; }
     inline bool atEof() { return !hasBufferedInput() && fileeof; }
     inline bool hasBufferedInput() { return buffered && (bufferedInput > bufferPosition); }
-    inline int  getHandle() { return fileHandle; }
+    inline uintptr_t getHandle() { return (uintptr_t)fileHandle; }
 
 protected:
     void   getStreamTypeInfo();
